+++ ChangeLog.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- Revised `Get-PSReminderDBInformation` to include Author and Comment from the Metadata table.
+
+### Fixed
+- Fixed parameter layout in help.
 
 ## [0.2.0] - 2024-07-19